Added room check

+++ controllers/versions/delete.ts
@@ -2,6 +2,11 @@
 
 export = (socket: SocketIO.Socket, doc: number, name: string, fn: Function) => {
 
+    if (Object.keys(socket.rooms).indexOf(''+doc) == -1) {
+        fn(true);
+        return;
+    }
+
     let sql: string = `
         DELETE FROM document_versions WHERE doc_id IN (
             SELECT doc_id FROM documents WHERE (doc_id = ? AND user_id = ?)
